requirements.txt
gspread==6.2.1
matplotlib==3.10.7
openai==2.8.1
Pillow
protobuf==3.20.3
streamlit==1.25.0
streamlit_drawable_canvas==0.9.3
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib-metadata==6.11.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.26.4
oauthlib==3.3.1
openai==2.8.1
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.5.0
platformdirs==4.5.0
plotly==6.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==311
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.29.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
streamlit==1.25.0
streamlit-drawable-canvas==0.9.3
streamlit-image-coordinates
streamlit-plotly-events==0.0.6
tenacity==8.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==4.3.1
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
yarg==0.1.9
zipp==3.23.0
